Fixed another small error in dataset and Small update to Corpus Cleaning
