+++ .idea/encodings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/Story_Database.csv" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/File Cleaning/Story_Database.csv" charset="UTF-8" />
   </component>
 </project>
